+++ src/main/java/online/kingdomkeys/kingdomkeys/client/ClientSetup.java
@@ -29,6 +29,7 @@ import online.kingdomkeys.kingdomkeys.client.render.*;
 import online.kingdomkeys.kingdomkeys.config.ModConfigs;
 import online.kingdomkeys.kingdomkeys.data.GlobalData;
 import online.kingdomkeys.kingdomkeys.data.PlayerData;
+import online.kingdomkeys.kingdomkeys.effects.ModMobEffects;
 import online.kingdomkeys.kingdomkeys.entity.ModEntities;
 import online.kingdomkeys.kingdomkeys.handler.ClientEvents;
 import online.kingdomkeys.kingdomkeys.handler.InputHandler;
@@ -242,7 +243,7 @@ public class ClientSetup {
 		if(!Utils.shouldRenderOverlay(player)) { //If it shouldn't render cause it's set to HIDE or WEAPON and not holding one
 			event.setCanceled(o.equals(COMMAND_MENU) || o.equals(MP_BAR) || o.equals(DRIVE_BAR) || o.equals(SHOTLOCK)); //Remove all these 4 bars
 			if(o.equals(HP_BAR) || o.equals(PLAYER_PORTRAIT)) { //Allow HP to be shown if KO'd
-				event.setCanceled(!globalData.isKO());
+				event.setCanceled(!player.hasEffect(ModMobEffects.KO));
 			}
 		} else { //If mode is set to SHOW or WEAPON while holding one
 			if(o.equals(MP_BAR)) { //Remove MP Bar is magics map is empty