+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/KOGui.java
@@ -9,6 +9,7 @@ import net.minecraft.client.gui.screens.TitleScreen;
 import net.minecraft.network.chat.Component;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.buttons.MenuButton;
 import online.kingdomkeys.kingdomkeys.data.GlobalData;
+import online.kingdomkeys.kingdomkeys.effects.ModMobEffects;
 import online.kingdomkeys.kingdomkeys.lib.Strings;
 import online.kingdomkeys.kingdomkeys.network.PacketHandler;
 import online.kingdomkeys.kingdomkeys.network.cts.CSGiveUpKO;
@@ -28,7 +29,6 @@ public class KOGui extends ChatScreen {
 			case "giveup" -> {
 				PacketHandler.sendToServer(new CSGiveUpKO());
 				this.minecraft.setScreen(null);
-				GlobalData.get(minecraft.player).setKO(false);
 			}
 			case "exit" -> {
 				if (this.minecraft.level != null) {
@@ -64,11 +64,8 @@ public class KOGui extends ChatScreen {
 			if(minecraft.player.getHealth() <= 0) {
 				onClose();
 			}
-			GlobalData globalData = GlobalData.get(minecraft.player);
-			if(globalData != null){
-				if(!globalData.isKO()){
-					onClose();
-				}
+			if(!minecraft.player.hasEffect(ModMobEffects.KO)){
+				onClose();
 			}
 		}
 		super.render(gui, p_render_1_, p_render_2_, p_render_3_);