+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/StopGui.java
@@ -34,15 +34,10 @@ public class StopGui extends Screen {
 	@Override
 	public void render(@NotNull GuiGraphics gui, int p_render_1_, int p_render_2_, float p_render_3_) {
 		if(minecraft.player != null) {
-
-			if(GlobalData.get(minecraft.player) != null) {
-	            //InputConstants.grabOrReleaseMouse(this.minecraft.getWindow().getWindow(), 212993, 0, 0);
-				if(minecraft.player.hasEffect(ModMobEffects.STOP)) {
-					onClose();
-				}
-				
+			if(minecraft.player.hasEffect(ModMobEffects.STOP)) {
+				onClose();
 			}
+
 		}
-		//super.render(matrixStack, p_render_1_, p_render_2_, p_render_3_);
 	}
 }