+++ src/main/java/online/kingdomkeys/kingdomkeys/effects/ModMobEffects.java
@@ -17,5 +17,7 @@ public class ModMobEffects {
             FREEZE = MOB_EFFECTS.register("freeze", () -> new FreezeEffect(MobEffectCategory.HARMFUL, 0xFFFFFF)),
             STOP = MOB_EFFECTS.register("stop", () -> new StopEffect(MobEffectCategory.HARMFUL, 0xB3408D)),
             GRAVITY = MOB_EFFECTS.register("gravity", () -> new GravityEffect(MobEffectCategory.HARMFUL, 0x63337F)),
-            AERO = MOB_EFFECTS.register("aero", () -> new AeroEffect(MobEffectCategory.BENEFICIAL, 0x60B86E));
+            AERO = MOB_EFFECTS.register("aero", () -> new AeroEffect(MobEffectCategory.BENEFICIAL, 0x60B86E)),
+            KO = MOB_EFFECTS.register("ko", () -> new KOEffect(MobEffectCategory.NEUTRAL, 0x60B86E));
+
 }