+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 **Added:**
 * Tags for each ore: "c:ores/blazing" etc.
+* Blizzard freeze effect, duration is longer for higher level magic and will stack duration for every hit
+* Aero, Gravity and Stop now work with mob effects.
+
+**Changed:**
+* KO system is now a mob effect, hopefully should be more stable than before.
+* KO effect will remove the lock on state.
 
 **Fixed:**
-* Proof of heart crashing the game.
-* Arrowgun and Wisdom form projectile rotation.
 * Gula boots crafting recipe
 * Potential fix for players being stuck in glide pos if going into a drive form.
+* Command Menu target selector will now only show party members within range.
+* Modern UI fix for level up message.